requirements.txt
transformers
peft
datasets
accelerate
scikit-learn
bitsandbytes
flash-attn
huggingface-hub
evaluate
pandas
pyyaml
wandb
requests
openai
tqdm
fasttext
